+++ libsa4py/cst_transformers.py
@@ -1,7 +1,10 @@
 from typing import Union, Dict, Tuple, List, Optional
+from collections import Counter
+from libsa4py.nl_preprocessing import NLPreprocessor
 import libcst as cst
 import libcst.matchers as match
 import re
+import regex
 
 
 class CommentAndDocStringRemover(cst.CSTTransformer):
@@ -846,3 +849,154 @@ class ParametricTypeDepthReducer(cst.CSTTransformer):
             return updated_node.with_changes(slice=cst.Index(value=cst.Name(value='Any')))
         else:
             return updated_node
+
+
+class TypeApplier(cst.CSTTransformer):
+    """
+    It applies (inferred) type annotations to a source code file.
+    Currently, it applies only variables' type annotations.
+    """
+
+    METADATA_DEPENDENCIES = (cst.metadata.ScopeProvider,)
+
+    def __init__(self, f_processeed_dict: dict):
+        self.f_processed_dict = f_processeed_dict
+        self.cls_visited = []
+        self.fn_visited = []
+
+        self.last_visited_assign_t_name = None
+        self.last_visited_assign_t_count = 0
+
+        self.current_cls_vars = None
+        self.current_fn_vars = None
+
+        self.nlp_p = NLPreprocessor()
+
+    def __get_fn_vars(self, var_name: str) -> dict:
+        for fn in self.f_processed_dict['funcs']:
+            if fn['name'] == self.nlp_p.process_identifier(self.fn_visited[-1]):
+                if var_name in fn['variables']:
+                    return fn['variables'][var_name]
+                else:
+                    continue
+
+    def __get_cls_vars(self, var_name: str) -> dict:
+        for cls in self.f_processed_dict['classes']:
+            if cls['name'] == self.cls_visited[-1]:
+                if var_name in cls['variables']:
+                    return cls['variables'][var_name]
+                else:
+                    continue
+
+    def __get_cls_fn_vars(self, var_name: str):
+        for cls in self.f_processed_dict['classes']:
+            if cls['name'] == self.cls_visited[-1]:
+                for fn in cls['funcs']:
+                    if fn['name'] == self.nlp_p.process_identifier(self.fn_visited[-1]):
+                        if var_name in fn['variables']:
+                            return fn['variables'][var_name]
+                        else:
+                            continue
+
+    def __get_mod_vars(self):
+        return self.f_processed_dict['variables']
+
+    def __get_var_names_counter(self, node, scope):
+        vars_name = match.extractall(node, match.AssignTarget(target=match.SaveMatchedNode(
+            match.Name(value=match.DoNotCare()), "name")))
+        return Counter([n['name'].value for n in vars_name if isinstance(self.get_metadata(cst.metadata.ScopeProvider,
+                                                                                           n['name']), scope)])
+
+    def visit_ClassDef(self, node: cst.ClassDef):
+        self.cls_visited.append(node.name.value)
+        self.current_cls_vars = self.__get_var_names_counter(node, cst.metadata.ClassScope)
+
+    def visit_FunctionDef(self, node: "FunctionDef"):
+        self.fn_visited.append(node.name.value)
+        self.current_fn_vars = self.__get_var_names_counter(node, cst.metadata.FunctionScope)
+
+    def leave_ClassDef(self, original_node: cst.ClassDef, updated_node: cst.ClassDef):
+        self.cls_visited.pop()
+        return updated_node
+
+    def leave_FunctionDef(self, original_node: cst.FunctionDef, updated_node: cst.FunctionDef):
+        self.fn_visited.pop()
+        return updated_node
+
+    def leave_SimpleStatementLine(self, original_node: cst.SimpleStatementLine,
+                                  updated_node: cst.SimpleStatementLine):
+
+        if match.matches(original_node, match.SimpleStatementLine(body=[match.Assign(targets=[match.AssignTarget(
+                target=match.Name(value=match.DoNotCare()))])])):
+
+            t: str = None
+            if len(self.cls_visited) != 0:
+                if len(self.fn_visited) != 0:
+                    # A class method's variable
+                    if self.current_fn_vars[original_node.body[0].targets[0].target.value] == self.last_visited_assign_t_count:
+                        t = self.__get_cls_fn_vars(self.nlp_p.process_identifier(original_node.body[0].targets[0].target.value))
+                else:
+                    # A class variable
+                    if self.current_cls_vars[original_node.body[0].targets[0].target.value] == self.last_visited_assign_t_count:
+                        t = self.__get_cls_vars(self.nlp_p.process_identifier(original_node.body[0].targets[0].target.value))
+            elif len(self.fn_visited) != 0:
+                # A module function's variable
+                if self.current_fn_vars[original_node.body[0].targets[0].target.value] == self.last_visited_assign_t_count:
+                    t = self.__get_fn_vars(self.nlp_p.process_identifier(original_node.body[0].targets[0].target.value))
+            else:
+                # A module's variables
+                t = self.__get_mod_vars()[self.nlp_p.process_identifier(original_node.body[0].targets[0].target.value)]
+
+            if t is not None:
+                t_annot_node = self.__name2annotation(self.resolve_type_alias(t))
+                if t_annot_node is not None:
+                    return updated_node.with_changes(body=[cst.AnnAssign(
+                        target=original_node.body[0].targets[0].target,
+                        value=original_node.body[0].value,
+                        annotation=t_annot_node,
+                        equal=cst.AssignEqual(whitespace_after=original_node.body[0].targets[0].whitespace_after_equal,
+                                            whitespace_before=original_node.body[0].targets[0].whitespace_before_equal))]
+                    )
+
+        return original_node
+
+    def __name2annotation(self, type_name: str):
+        ext_annot = lambda t: match.extract(cst.parse_module("x: %s=None" % t).body[0].body[0],
+                                            match.AnnAssign(target=match.Name(value=match.DoNotCare()),
+                                                            annotation=match.SaveMatchedNode(match.DoNotCare(), "type")))['type']
+        try:
+            return ext_annot(type_name)
+        except cst._exceptions.ParserSyntaxError:
+            return None
+
+    def visit_AssignTarget(self, node: cst.AssignTarget):
+        if match.matches(node, match.AssignTarget(target=match.Name(value=match.DoNotCare()))):
+            if self.last_visited_assign_t_name == node.target.value:
+                self.last_visited_assign_t_count += 1
+            elif self.last_visited_assign_t_count == 0:
+                self.last_visited_assign_t_count = 1
+            else:
+                self.last_visited_assign_t_count = 1
+            self.last_visited_assign_t_name = node.target.value
+
+    def resolve_type_alias(self, t: str):
+        type_aliases = {'^{}$|^Dict$|(?<=.*)Dict\[\](?<=.*)|(?<=.*)Dict\[Any, *?Any\](?=.*)|^Dict\[unknown, *Any\]$': 'dict',
+                        '^Set$|(?<=.*)Set\[\](?<=.*)|^Set\[Any\]$': 'set',
+                        '^Tuple$|(?<=.*)Tuple\[\](?<=.*)|^Tuple\[Any\]$|(?<=.*)Tuple\[Any, *?\.\.\.\](?=.*)|^Tuple\[unknown, *?unknown\]$|^Tuple\[unknown, *?Any\]$|(?<=.*)tuple\[\](?<=.*)': 'tuple',
+                        '^Tuple\[(.+), *?\.\.\.\]$': r'Tuple[\1]',
+                        '\\bText\\b': 'str',
+                        '^\[\]$|(?<=.*)List\[\](?<=.*)|^List\[Any\]$|^List$': 'list',
+                        '^\[{}\]$': 'List[dict]',
+                        '(?<=.*)Literal\[\'.*?\'\](?=.*)': 'Literal',
+                        '(?<=.*)Literal\[\d+\](?=.*)': 'Literal',  # Maybe int?!
+                        '^Callable\[\.\.\., *?Any\]$|^Callable\[\[Any\], *?Any\]$|^Callable[[Named(x, Any)], Any]$': 'Callable',
+                        '^Iterator[Any]$': 'Iterator',
+                        '^OrderedDict[Any, *?Any]$': 'OrderedDict',
+                        '^Counter[Any]$': 'Counter',
+                        '(?<=.*)Match[Any](?<=.*)': 'Match',
+                        '^\.(.+)': r'\1',
+                        '(?<=.*)Optional\[\](?<=.*)': 'Optional'}
+        for t_alias in type_aliases:
+            if regex.search(regex.compile(t_alias), t):
+                t = regex.sub(regex.compile(t_alias), type_aliases[t_alias], t)
+        return t